% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot motor vehicle fatalities on a map.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer value identifying a particular US state}

\item{year}{An integer value representing a year.}
}
\value{
This function returns a map (plot)
}
\description{
This function  plots motor vehicle fatalities for a specific state of the USA and year (\code{state.num}, \code{year}) and plots them on a map.
If no fatalities occured in a given state/year combination the user is shown a message which explains this.
}
\note{
This function has dependencies on \code{\link[maps]{map}} and \code{\link[graphics]{points}}

This function will return an error if an invalid state code is provided as input.
}
\examples{
  fars_map_state(1,2014)
  fars_map_state(12,2015)

}
