% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a file into a dataframe}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string representing a filename.}
}
\value{
This function returns a dataframe containing the data read from the specified file.
  As a side effect, this function also prints the data to the console.
}
\description{
This function checks that a specified file (\code{filename}) exists and if so, reads the data from the file into a dataframe.
}
\note{
This function has dependencies on \code{\link[readr]{read_csv}} and \code{\link[dplyr]{tbl_df}}

If the filename supplied does not exist in the user's working direcrtory then this funtion returns an error.
}
\examples{
  fars_read(newfile.txt)
  fars_read(oldfile.csv)

}
