% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read a file or files into a list of dataframes.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An integer vector of years}
}
\value{
This function returns a list of dataframes. It will error if an invalid year value is supplied.
}
\description{
This function reads in a vector of filename(s) (\code{years}) and generates a dataframe for each one.
  The dataframes consist of two columns, MONTH and year. The dataframes have the same number of rows as the source file.
  Each dataframe is held as an item in a list.
}
\note{
This function has dependencies on \code{\link[dplyr]{mutate}} and \code{\link[dplyr]{select}}

If the user supplies a year for which no data exists then this function returns an error.
}
\examples{
 \dontrun{
  fars_read_years(c(2013,2014,2015))

  yrs <- c(2013,2014,2015)
  fars_read_years(yrs)
}

}
