% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarise motor vehicle fatality information.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An integer vector of years}
}
\value{
This function returns a dataframe containing counts of motor vehicle fatalities by month and year.
}
\description{
This function produces a table of counts of motor vehicle fatalities by year and month, for a vector of years provided by the user (\code{years})
}
\details{
importFrom dplyr bind_rows group_by summarize
importFrom tidyr spread
}
\note{
This function has dependencies on \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarize}} and \code{\link[tidyr]{spread}}

This function will return an error if the user provides a year for which data does not exist as input.
}
\examples{
  df <- fars_summarize_years(c(2013,2014,2015))

  yrs <- c(2013,2014,2015)
  fars_summarize_years(yrs)

}
