% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Dynamically generate a file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer value representing a year.}
}
\value{
This function returns a file name as a string.
}
\description{
This function creates a filename based on a year (\code{year}) input by the user.
}
\note{
If the user does not supply an input which can be treated as an integer then this function returns an error.
}
\examples{
 \dontrun{
  make_filename(2015)
  make_filename(2016)
}

}
