% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Maps FARS crash locations by state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Integer state code. See \url{https://crashstats.nhtsa.dot.gov/Api/Public/Publication/812316} (PDF), p. 24 for a list.}

\item{year}{Integer year.}
}
\description{
Plots a map showing FARS crash locations for a specified state and year.
}
\details{
The \code{maps} package must be loaded so that the the \code{stateMapEnv} object is
available. (The \code{maps} package does not export \code{stateMapEnv}.) Do this by
running \code{library(maps)} .

Crashes with \code{LOGITUD > 900} or \code{LATITUDE > 90} are excluded.

Invalid \code{state} or \code{year} values will produce errors.
}
\examples{
fars_map_state(19, 2015) # 19 is Iowa

}

