% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number ID associated with the state of interest. This 
argument is cast as an integer inside the function, and so can be safely 
passed as a character, integer, or numeric quantity without causing errors.}

\item{year}{The year for which the state plot is desired. This argument is 
cast as an integer inside the function, and so can be safely passed as a
character, integer, or numeric quantity without causing errors.}
}
\value{
This function creates a plot of the United States, zoomed in on the
   specified state, with the location of individual fatal collisions plotted 
   as points.
}
\description{
This function plots the location of fatal collisions recorded by FARS for a 
given state and year. The collision locations are drawn over a map of the 
United States cropped to contain the latitude & longitude of the entries for 
the specified state.
}
\details{
If the state number given does not exist in the FARS data for the specified
year, an error will be raised. An invalid year will lead to an error being 
raised when attempting to read in the data with 
\code{data <- fars_read(filename)}. This function replaces unphysical latitude 
and longitude values with NA; note that it does not verify that the 
coordinates of each collision fall within the boundaries of the specified 
state. Any collisions with missing latitude or longitude are not plotted.
}
\examples{
\dontrun{fars_map_state(1, 2013)}
\dontrun{fars_map_state("2", "2014")}
\dontrun{fars_map_state(0, 2015)} # Invalid state.num
\dontrun{fars_map_state(10,2010)} # Invalid year (outside of 2013-2015)

}
