% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot the location of fars accident reports}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the numeric code for the state or territory to be plotted.}

\item{year}{A 4-digit year number (YYYY) specifying the FARS file to be read.}
}
\value{
The function returns a NULL value.
}
\description{
This function plots the accident reports of a given year
and in a given state on a map of the state.
}
\details{
The function will fail if the state code is not in FARS file.
If there are no accidents to plot, a message is produced to that effect.

The following functions are imported: \code{maps::map}, \code{graphics::points}, \code{dply::filter}.
}
\examples{
fars_map_state(1, 2014)
# following will fail
fars_map_state(1, 2019)

}
