% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{FARS State Accident Plotter}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state's number as an integer}

\item{year}{The desired year in YYYY format as an integer}
}
\value{
Function returns a map of the given state with accident
reports that have valid longitude and latitute points plotted on the
map.
}
\description{
This function  plots the accidents of a given year and state to the
state's map. It uses fars_read and make_filename functions included
in this package. It also imports numerous functions from dplyr, maps
and graphics packages for its' function
}
\details{
This function requires a valid state number and  a year that
is included in FARS dataset to operate. The working directory needs to
include .csv.bz2 files included in the data folder of fars_data.zip
file.
}
\examples{
\dontrun{
fars_map_state(25,2014)
fars_map_state(1,2015)
fars_map_state(6,2013)
}
}
