% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads data from a FARS data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the file to read.}
}
\value{
Returns a data frame containing the FARS data.
}
\description{
Reads data from a FARS data file.
}
\details{
Generates an error if the file does not exist.
}
\examples{
filename <- system.file("extdata", make_filename(2015), package = "FARSAssignment")
data <- fars_read(filename)
}

