% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{String containing the input data file name.}
}
\value{
This function returns a tibble containing the data from the specified
   file.
}
\description{
This function quietly loads data from a specified file. If the given 
file does not exist, an error will be raised. It has been designed to be used
in conjunction with \code{make_filename(year)}.
}
\examples{
\dontrun{fars_read("accident_2013.csv.bz2")}

}
