% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a FARS Data File}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the .csv format file to be read.}
}
\value{
This function returns a \code{tbl_df} form of a \code{data.frame}.
}
\description{
This function reads a (possibly compressed) .csv format file specified by the \code{filename} argument and returns it as a
\code{tbl_df data.frame}.
}
\details{
The function will fail if the file specified by the \code{filename} argument does not exist. 
The \code{read_csv} function may fail if the file is not readable. Any warnings introduced by parsing failures are suppressed.

The following functions are imported: \code{readr::read_csv}, \code{dplr::tbl_df}.
}
\examples{
#improper name format
fars_read("fars.csv")

fars_read("accident_2014.csv.bz2")


}
