% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Fars Data Import}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{this parameter takes the expected name of the file
that contains Fatality Analysis Reporting System data}
}
\value{
This function return a data frame using dplyr package's
tbl_df function and readr package's read_csv function
}
\description{
This simple function serves as a basic check and data import for the
several other functions provided in this package
}
\details{
this function requires you to unzip the fars_data file in
order to reach the .bz2 data files that is contained inside for
each individual year. Working directory needs to contain the files
that are included in data folder of fars_data .zip file
}
\examples{
\dontrun{
fars_data(accident_2014.csv.bz2)
}
}
