% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of years for which condensed Month-Year data is desired.
Each year may be specified as a character, integer, or numeric quantity
without causing any errors (assuming the year itself is valid).}
}
\value{
This function returns a list of tibbles, one for each year in the 
input list. Output for any invalid years will be NULL.
}
\description{
This function condenses the FARS data for a list of years down to the month
and year of each fatal injury due to a motor vehicle collision. Note that a 
new "year" column is added to the table, although FARS already contains a 
"YEAR" column. If one of the years in the input list is invalid, NULL will
be returned for that year and a warning will be raised. For the available 
data sets, valid years are 2013, 2014, and 2015.
}
\examples{
\dontrun{fars_read_years(c(2013, 2014, 2015))}
\dontrun{fars_read_years(2010:2020)} # Note that some years in this range
                                     # are invalid
                             
}
