% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{FARS Data Reader}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{this parameter takes a vector of numbers that represent
the desired years as an input.}
}
\value{
the function returns a table that provides month and year of
each reported incident.
}
\description{
This function retrieves the reported incidents from requested data
files and creates a table that provides month and year information
the incident has occured. It also uses fars_read and make_filename
functions contained in this package, and imports dplyr package.
}
\details{
This function expects to get the years as numbers in YYYY
format and have the relevant .csv.bz2 files in the working directory.
The function sends a warning message for each year file it wasn't
able to locate.
}
\examples{
\dontrun{
fars_read_years (c(2013,2015))
fars_read_years (c(2014,2015)) }
}
