% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of years for which monthly total fatalities will be 
tabulated. Each year may be specified as a character, integer, or numeric
quantity without causing any errors (assuming the year itself is valid).}
}
\value{
This function returns a summary table of total monthly fatalities for
   each year in the input list. Months remain in a column and each year has
   its own column (as a result of \code{tidyr::spread(year, n)}). The values
   in the table represent total monthly fatalities for a given month-year
   combination.
}
\description{
This function summarizes the number of fatal injuries by month and year in 
the FARS data set for all of the input years. Valid years (given the data 
files that are currently available) are 2013, 2014, and 2015. Any invalid 
years will be excluded from the output summary. An error will be thrown by
\code{fars_read_years()} for any invalid years.
}
\examples{
\dontrun{fars_summarize_years(c(2013, 2014, 2015))}
\dontrun{fars_summarize_years(2013:2015)} 

}
