% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{FARS Data Summarizer}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{this parameter takes a vector of numbers that represent
the desired years as an input.}
}
\value{
the function returns a table that provides monthly number of
incidents for each year.
}
\description{
This function retreves the data for requested years and summarizes
the data in a table for each month.It uses fars_read and make_filename
and fars_read_years functions contained in this package. Function also
uses dplyr and tidyr packages.
}
\details{
This function expects to get the years as numbers in YYYY
format and have the relevant .csv.bz2 files in the working directory.
The function sends a warning message for each year file it wasn't
able to locate.
}
\examples{
\dontrun{
fars_summarize_years (c(2013,2015))
fars_summarize_years (c(2014,2015))}
}
