% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Makes a FARS filename for a specified year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer specifying the year.}
}
\value{
A character string containing the filename.
}
\description{
Makes a FARS filename for a specified year
}
\examples{
filename <- make_filename(2015)

}

