% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year for which a filename is needed. This argument is cast
as an integer inside the function, and so can be safely passed as a 
character, integer, or numeric quantity without causing any errors.}
}
\value{
This function returns a string containing the filename for FARS data
   for the input year.
}
\description{
This function creates the appropriate filename string for the FARS data for 
a given year. The format of the filename is "accident_YYYY.csv.bz2". This 
function does not check if this year is valid.
}
\examples{
\dontrun{make_filename(2013)}
\dontrun{make_filename("2014")}

}
