% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a FARS Data File Name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The 4-digit year number (YYYY) to be embedded in the file name.}
}
\value{
This function returns a \code{string} of the form: "accident_YYYY.csv.bz2".
}
\description{
This function produces a filename which follows the conventions of 
FARS (Federal Accident Reports) naming including the supplied \code{year} argument.
}
\examples{
make_filename(2014)

}
