% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Expected Filename Creator}
\usage{
make_filename(year)
}
\arguments{
\item{year}{this parameter takes a number as integer and in order
to create that year's expected filename}
}
\value{
Function returns a string that should be same as the name
of the desired datafile
}
\description{
This function takes a number as an input and creates the predicted
filename for the data file that should contain the corresponding
year's Fatality Analysis Reporting System data
}
\details{
Function expects to get number of a year in the YYYY format.
 Even though it can coerce double numbers to integers and accept
 invalid numbers as an input, any number that is not one of the years
 FARS data contains  would result in an error. The function also
 operates with the asumption that the file names contained in the data
 folders are not changed.
}
\examples{
\dontrun{
make_filename(2014)
make_filename(2015) }
}
